+++ crates/icy_board_engine/src/vm/tests/mod.rs
@@ -7,6 +7,7 @@
 
 #![cfg(test)]
 
+mod arrays;
 mod board_session;
 mod builtin_enums;
 mod contacts;