+++ crates/ppl-lsp/src/documentation.rs
@@ -432,6 +432,8 @@ pub fn get_function_hover(func: &FunctionDefinition) -> Option<Hover> {
         FuncOpCode::AreaId => get_sig_hint(sig, fl!(crate::LANGUAGE_LOADER, "hint-function-areaid")),
         FuncOpCode::WebRequest => get_sig_hint(sig, fl!(crate::LANGUAGE_LOADER, "hint-function-webrequest")),
         FuncOpCode::Len_Dim => get_sig_hint(sig, fl!(crate::LANGUAGE_LOADER, "hint-function-len_dim")),
+        FuncOpCode::ElementCount => get_sig_hint(sig, fl!(crate::LANGUAGE_LOADER, "hint-function-element_count")),
+        FuncOpCode::ElementAt => get_sig_hint(sig, fl!(crate::LANGUAGE_LOADER, "hint-function-element_at")),
         FuncOpCode::BASE64ENC => get_sig_hint(sig, fl!(crate::LANGUAGE_LOADER, "hint-function-base64enc")),
         FuncOpCode::BASE64DEC => get_sig_hint(sig, fl!(crate::LANGUAGE_LOADER, "hint-function-base64dec")),
         FuncOpCode::SHA256 => get_sig_hint(sig, fl!(crate::LANGUAGE_LOADER, "hint-function-sha256")),